Amazon Review Sentiment Analysis
The goal of this project is to predict the rating of an Amazon product by analyzing the sentiment of the review given by the customer. The model will be used to predict reviews provided by customers on different portals which in turn will be input data to manage the brand and reputation of the product. Both rule based and machine learning sentiment analysis models were evaluated on the data. For Rule based sentiment analyzer Vader and TextBlob were evaluated whereas for Machine Learning based analyzer Logistic regression, Naive Bayes and Random Forest were explored.